refactor: moved dummy factoids as well
